MoviePage 메서드 서비스로 이동
