+++ review/src/main/java/movieReview/review/controller/SearchFunc/MoviePage.java
@@ -29,7 +29,6 @@ import java.util.List;
 public class MoviePage {
     private final UploadService uploadService;
     private final getMovieInfoService getMovieInfoService;
-    private final photoUriInfo photoUriinfo;
 
     @ResponseStatus(HttpStatus.OK)
     @PostMapping("/viewOneMovie")
@@ -39,7 +38,7 @@ public class MoviePage {
         List<JpaMovieInfo> uri = new ArrayList<>();
         List<JpaMovieInfo> movie = getMovieInfoService.getMovie(movieinfo);
         for (int i = 0; i < movie.size(); i++) {
-            uri.add(makeUri(movie.get(i)));
+            uri.add(uploadService.makeUri(movie.get(i)));
             jpaMovieInfos.add(uri.get(i));
         }
 
@@ -59,18 +58,5 @@ public class MoviePage {
         return "/SerachPage/MovieSerachPage";
     }
 
-
-    public JpaMovieInfo makeUri(JpaMovieInfo movie) {
-        photoUriinfo.setPhotoUri(movie.getPhotoUri());
-        photoUriinfo.setPhotoOriName(movie.getPhotoOriName());
-
-        photoUriInfo photoUriInfo = uploadService.showPhoto(photoUriinfo);
-        ClassPathResource resource = new ClassPathResource("/moviePhoto/" + photoUriInfo.getPhotoOriName());
-
-        Path path1 = Paths.get(resource.getPath());
-
-        movie.setPhotoUri(path1.toString());
-        return movie;
-    }
 
 }
